feat: delete record in idb after sync on chain
