+++ providers/ic/src/providers/data/doc.providers.ts
@@ -1,10 +1,14 @@
 import {DeleteDoc, Doc, DocData, DocEntries, SnapshotDoc} from '@deckdeckgo/editor';
-import {deleteData, entries} from '../../api/data.api';
+import {entries} from '../../api/data.api';
+import {deleteData} from '../../services/data.services';
 
 export const docEntries: DocEntries = async (_userId: string): Promise<Doc[]> =>
   entries<DocData>({startsWith: '/docs/', notContains: '/paragraphs/'});
 
-export const deleteDoc: DeleteDoc = async (docId: string, updated_at?: Date | number | BigInt): Promise<void> =>
+export const deleteDoc: DeleteDoc = async (
+  docId: string,
+  updated_at?: Date | number | BigInt
+): Promise<void> =>
   deleteData({key: `/docs/${docId}`, ...(updated_at !== undefined && {id: docId, updated_at})});
 
 // Backwards compatibility with current publish mask in studio and Firebase support. In case of IC we actually do not need a snapshot, publish is synchronous on the client side.
