+++ providers/ic/src/services/data.services.ts
@@ -1,7 +1,7 @@
 import type {DataRecord} from '@deckdeckgo/editor';
-import {update} from 'idb-keyval';
-import {setData as setDataApi} from '../api/data.api';
-import {_SERVICE as DataBucketActor} from '../canisters/data/data.did';
+import {update, del} from 'idb-keyval';
+import {deleteData as deleteDataApi, setData as setDataApi} from '../api/data.api';
+import {DelData, _SERVICE as DataBucketActor} from '../canisters/data/data.did';
 import {LogWindow} from '../types/sync.window';
 
 export const setData = async <D>({
@@ -32,4 +32,35 @@ export const setData = async <D>({
   log?.({msg: `[set][done] ${key}`, duration: t1 - t0});
 
   return updatedEntity;
+};
+
+export const deleteData = async ({
+  key,
+  actor,
+  log,
+  data
+}: {
+  key: string | undefined;
+  actor?: DataBucketActor;
+  log?: LogWindow;
+  data?: DelData;
+}): Promise<void> => {
+  if (!key) {
+    // Should never happen but, you never know
+    return;
+  }
+
+  log?.({msg: `[delete][start] ${key}`});
+  const t0 = performance.now();
+
+  await deleteDataApi({key, actor, data});
+
+  // TODO: remove if once deprecated code will be removed
+  if (data !== undefined) {
+    // Paragraph has been deleted in the cloud, we can delete the local record
+    await del(key);
+  }
+
+  const t1 = performance.now();
+  log?.({msg: `[delete][done] ${key}`, duration: t1 - t0});
 };
